Refactor dashboard into smaller components
Split the monolithic dashboard/index.tsx (1,336 lines) into focused,
reusable components (now 162 lines):

- StreamHealthMonitor: Stream connection quality metrics
- ViewerEngagementScore: Engagement score calculation and display
- StreamGoalsTracker: Progress tracking for stream goals
- ActivityFeed: Filterable event feed with type filters
- QuickActionsPanel: Floating action button panel
- QuickStats: Summary stat cards grid
- RecentChat: Recent chat messages display
- RecentEvents: Recent stream events display
- RecentStreams: Recent livestreams list
- DashboardQuickActions: Navigation quick action cards
- DashboardSkeletons: Loading skeleton components

All components are exported from ~/components/dashboard for clean imports.
