+++ frontend/src/components/dashboard/index.ts
@@ -1,0 +1,18 @@
+export { default as ActivityFeed } from "./ActivityFeed";
+export { default as DashboardQuickActions } from "./DashboardQuickActions";
+export {
+	ActivityFeedSkeleton,
+	DashboardLoadingSkeleton,
+	QuickStatsSkeleton,
+	RecentChatSkeleton,
+	RecentEventsSkeleton,
+	RecentStreamsSkeleton,
+} from "./DashboardSkeletons";
+export { default as QuickActionsPanel } from "./QuickActionsPanel";
+export { default as QuickStats } from "./QuickStats";
+export { default as RecentChat } from "./RecentChat";
+export { default as RecentEvents } from "./RecentEvents";
+export { default as RecentStreams } from "./RecentStreams";
+export { default as StreamGoalsTracker } from "./StreamGoalsTracker";
+export { default as StreamHealthMonitor } from "./StreamHealthMonitor";
+export { default as ViewerEngagementScore } from "./ViewerEngagementScore";
